# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST for LiveCart SaaS — live commerce order detection, cart consolidation, and integrations.",
        "title": "LiveCart API",
        "termsOfService": "https://livecart.com/terms",
        "contact": {
            "name": "LiveCart Support",
            "email": "support@livecart.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's profile including store information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_user.GetMeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/stores": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all stores the authenticated user belongs to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all stores for current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_user.GetUserStoresResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/users/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new user and store if not exists, or returns existing user. Call this after first sign-in.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Sync user on first access",
                "parameters": [
                    {
                        "description": "Optional store information for new users",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/apps_api_internal_user.SyncUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_user.SyncUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_user.SyncUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.ValidationEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apps_api_internal_user.GetMeResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storeId": {
                    "type": "string"
                },
                "storeName": {
                    "type": "string"
                },
                "storeSlug": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "apps_api_internal_user.GetUserStoresResponse": {
            "type": "object",
            "properties": {
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apps_api_internal_user.UserStoreResponse"
                    }
                }
            }
        },
        "apps_api_internal_user.SyncUserRequest": {
            "type": "object",
            "properties": {
                "storeName": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "storeSlug": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
        "apps_api_internal_user.SyncUserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "onboardingComplete": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storeId": {
                    "type": "string"
                },
                "storeName": {
                    "type": "string"
                },
                "storeSlug": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "apps_api_internal_user.UserStoreResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storeId": {
                    "type": "string"
                },
                "storeName": {
                    "type": "string"
                },
                "storeSlug": {
                    "type": "string"
                }
            }
        },
...
        "livecart_apps_api_lib_httpx.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "livecart_apps_api_lib_httpx.ValidationEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}